// src/app/masonry/ngx-masonry-options.interface.ts
/**
 * Masonry Options used for the animation
 */
export interface NgxMasonryOptions {
    itemSelector?: string;
    columnWidth?: number | string;
    gutter?: number | string;
    percentPosition?: boolean;
    stamp?: string;
    fitWidth?: boolean;
    originLeft?: boolean;
    originTop?: boolean;
    containerStyle?: string;
    transitionDuration?: string;
    resize?: boolean;
    initLayout?: boolean;
}